Add coloured logo
This logo should be used as an app icon.
It features:
- the updated FT letters (they are now financier)
- a subtle gradient

**WARNING**
The current image service (v1) isn't able to convert this to a PNG successfully.
If you want to use this image dynamically falling back to a PNG for older browsers
please wait for v2 of the image service.

If you need a PNG version of this logo now, one is available here:
http://res.cloudinary.com/financial-times/image/upload/v1472668474/brand-fastft_wmqjtz.png

If you need it at a different size, please ask the origami team, or create your
own until v2 of the image service has shipped.