Changing the bounding box of brand-ft
Because this is used in a few places, we can't simply change the
bounding box of this icon without breaking some places where it is used.

Unfortunately this use is not really a valid use of the letters 'FT' as
the spacing in the box is part of the logo so we're a bit stuck.

We should look at if the uses of FT like this are considered valid by
the branding team. Not providing a this svg will mean people will work
around us by just using financier, but maybe there is another way to
do the thing they want to do that is more in keeping with the FT branding